+++ src/utils.js
@@ -1,7 +1,0 @@
-export const cn = className =>
-  typeof className === 'string'
-    ? className
-        .split('.')
-        .join(' ')
-        .trim()
-    : '';